Refactor firebase to organize functions better
